pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [
  { name = "Bertil Varenhorst", email = "bertil.varenhorst@ericsson.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "flask==3.0.3",
  "pytest==8.2.2",
  "pyquery==2.0.0",
  "ruff==0.4.8",
  "pre-commit==3.7.1",
  "mypy==1.10.0",
  "pytest-cov==5.0.0",
  "python-dotenv==1.0.1",
  "loguru==0.7.2",
]
description = ''
dynamic = ["version"]
keywords = []
license = "MIT"
name = "plantuml-gui"
readme = "README.md"
requires-python = ">=3.8"

[project.urls]
Documentation = "https://github.com/Bertil Varenhorst/plantuml-gui#readme"
Issues = "https://github.com/Bertil Varenhorst/plantuml-gui/issues"
Source = "https://github.com/Bertil Varenhorst/plantuml-gui"

[tool.hatch.version]
path = "src/plantuml_gui/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = ["mypy>=1.0.0"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/plantuml_gui tests}"

[tool.hatch.build.targets.wheel]
packages = ["src/plantuml_gui"]

[tool.coverage.run]
branch = false
omit = ["src/plantuml_gui/__about__.py"]
parallel = true

[tool.coverage.paths]
plantuml_gui = ["src/plantuml_gui"]
tests = ["tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "I", "N"]
